Using yypush_buffer_state in main and parse_files
Since I'm using Flex's internal stack for include files, it is then
necessary to use it in main for adding user specified files. Cannot use
yy_switch_to_buffer as it seems to mess up Flex's stack. I also should
not switch/push a file on the EOF action or on the call to yywrap, as it
is far more difficult to know when the file is being added.