Using E() macro to create tree with NULL data for %empty actions
Semantic actions for epsilon ought to have the same type (tree *) as all
other actions; throwing NULL into the mix just asked for
trouble. Instead, I now create an empty tree with NULL data, which can
be checked for safely.

Also renamed p() macro to P() for less chance of it causing a weird
conflict with a function ending in ...p();