Major refactor of information gathering and typeinfo creation
Added a generic tree traversal to find the first token (in a pre-order
search) with the given target category, or the 'before' category if
encountered first.

Added helpers for getting identifier, pointer, array, class info using
this traversal.

Programmatically this is much cleaner and easier to understand, and
should also handle different tree shapes for rules a bit
better. However, this is a computationally most costly algorithm, since
each piece of info requires a recursive search on that subtree, rather
than just grabbing the information. But programmer time and maintenance
win over CPU cycles.

The typeinfo_new() now just allocates the struct, and pieces of
information are added on as they're found, rather than attempting to
find it all and send it to a variadic "constructor" function which was
way more headache than it was worth.

Note that standard libraries are currently no longer implemented, as
they require some refactoring due to the removal of the constructor.