Adding support for class constructors
Note that we are not supporting function overloading, so only one
constructor is allowed. At some point we will want to check if no
constructor was provided, and add a default.

Since the class is stored in the symbol table under its name as the key,
and class constructor functions share the name of their class, for now I
am storing it as classname_ctor, and will need to translate function
calls that use a classname to this suffixed key.