Using typeinfo_compare() during insertion for functions
Rather than typeinfo_list_compare() which is handled in the former
comparison, in addition to the return types of the function.

Fixing bug caused by typeinfo_new() assigning true to the function's
pointer flag because of a pointer existing before the identifier of the
function. There are better fixes, but 120++ does not have function
pointers.