Performing type checking in a second tree traversal
Easier to grok logically. Necessitates type_check() perform a pre-order
search of the entire syntax tree, type checking on expressions where
necessary, and noop for non-expressions. Avoids tricky business with
initializers and race to symbol insertion. Also necessitates handling of
nested scopes when encountering function definitions, but avoids
trickiness with implementing type checking concurrently with the
tree_preorder() application of handle_function().

Note that the INITIALIZER production rule must be explicilty ignored
when encountered during recursion of a declarator.

Type checking works for simple assignments of literal values and
symbols, as well as calls to the new operator.