Refactoring to use class_member() over get_class()
Stricter search for CLASS_NAME constrained to DIRECT_DECL(4|5) subtree,
which represents a class::name or class::class declaration, the latter
specifically in *definitions* of class member functions.

Prior use of get_class() required the first child node of a
FUNCTION_DEF(1|2) to exist (and be the return type) as it skipped over
it and explicitly checked the next node. With the addition of handling
constructor definitions, the former function definition also needed to
be handled, but lacked the child because ctors are a special case
with an immutable and unwritten return type.

Trying to add a faux type failed when the tree was freed because it the
insertion caused a double free.