Handling member function calls
Major caveat: this will break if attempting to access a public member
field, which are unsupported in 120++, because this assumes any field
access invokes a member function. For a member function call, the
instance is pushed as the first parameter (and it is derefenced first if
it is a pointer).

Does some funky stuff to get a full name reprensentation of the
function. At some point, when freeing the final code list, the ops
should free their name field.