Handling external C functions
If a function is not defined (that is, its symbol table is NULL), call
the function directly in TAC-C. This requires iterating backward to get
parameter addresses, but allows full C functionality. C headers will not
automatically pull-in prototypes, so a faux C header with common ones is
included, and others can simply be declared in source files. TODO: pass
C headers through to TAC-C.